refactor: 마이페이지 구조 리팩토링
- container-presenter 패턴 적용
- ui 컴포넌트 pages/view에서 관리
- view의 레이아웃 컴포넌트만 해당 컴포넌트에서 관리하고, 그 외 하위 컴포넌트는 components/MyPage에서 관리하도록 수정
