+++ frontend/src/components/common/PageTitle.tsx
@@ -7,7 +7,7 @@ const PageTitle = ({ title }: TitleProps) => {
     <h1 className={"text-bold-l mb-6 text-gray-black dark:text-white"}>
       {title}
     </h1>
-  )
-}
+  );
+};
 
 export default PageTitle;
