+++ frontend/src/components/mypage/ProfileIcon.tsx
