+++ frontend/src/components/questions/detail/QuestionList.tsx/index.tsx
@@ -1,4 +1,4 @@
-import { useGetQuestion } from "@/hooks/api/useGetQuestionContent";
+import { useGetQuestionContent } from "@/hooks/api/useGetQuestionContent";
 import QuestionItem from "./QuestionItem";
 
 const QuestionList = ({ questionId }: { questionId: string }) => {
@@ -7,7 +7,7 @@ const QuestionList = ({ questionId }: { questionId: string }) => {
     isLoading,
     isError,
     error,
-  } = useGetQuestion(questionId);
+  } = useGetQuestionContent(questionId);
 
   if (isLoading) return <div>로딩 중</div>;
   if (isError) return <div>에러가 발생했습니다: {error.message}</div>;
