+++ frontend/src/components/questions/detail/QuestionTitle.tsx/index.tsx
@@ -1,4 +1,4 @@
-import { useGetQuestion } from "@/hooks/api/useGetQuestionContent";
+import { useGetQuestionContent } from "@/hooks/api/useGetQuestionContent";
 import { MdEdit } from "react-icons/md";
 import { RiDeleteBin6Fill } from "react-icons/ri";
 import { FaRegBookmark } from "react-icons/fa";
@@ -9,7 +9,7 @@ const QuestionTitle = ({ questionId }: { questionId: string }) => {
     isLoading,
     isError,
     error,
-  } = useGetQuestion(questionId);
+  } = useGetQuestionContent(questionId);
 
   if (isLoading) return <div>로딩 중</div>;
   if (isError) return <div>에러가 발생했습니다: {error.message}</div>;
