+++ frontend/src/pages/MyPage/index.tsx
@@ -2,7 +2,7 @@ import { useEffect } from "react";
 import { useNavigate } from "react-router-dom";
 import useAuth from "@hooks/useAuth";
 import useToast from "@hooks/useToast";
-import MyPageView from "./MyPageView";
+import MyPageView from "./view";
 
 const MyPage = () => {
   const { isLoggedIn, nickname } = useAuth();
