+++ frontend/src/pages/MyPage/view/index.tsx
@@ -1,8 +1,8 @@
 import SidebarPageLayout from "@components/layout/SidebarPageLayout";
 import PageTitle from "@components/common/PageTitle";
-import ProfileEditModal from "@components/mypage/ProfileEditModal";
-import Profile from "@components/mypage/Profile";
-import QuestionSection from "@components/mypage/QuestionSection";
+import ProfileEditModal from "@/pages/MyPage/view/ProfileEditModal";
+import Profile from "@/pages/MyPage/view/Profile";
+import QuestionSection from "@/pages/MyPage/view/QuestionSection";
 
 interface MyPageViewProps {
   nickname: string;
@@ -17,7 +17,7 @@ const MyPageView = ({ nickname }: MyPageViewProps) => {
         <Profile nickname={nickname} />
         <QuestionSection />
       </div>
-    </SidebarPageLayout >
+    </SidebarPageLayout>
   );
 };
